Module Sc_ltest.Framac

1. Process command related functions

type funs = [

| `Auto of string

| `All

| `Only of string list

]

type options

module Cmd : Sc_sys.Ezcmd.T

Command line representation for frama-c

val lannot_cmd : (module Types.CONFIG) -> ?include_dirs:Sc_sys.File.TYPES.dir list -> cover_criterion:string -> annoted_functions:funs -> skipped_functions:string list -> output_file:string -> [> `C ] Sc_sys.File.TYPES.file -> Sc_sys.Process.command Lwt.t

lannot_cmd annot_kind proj Returns the command line for starting `lannot` with the coverage criterion `annot_kind` and the label file generated. You can get the list of all coverage criterions with `$ frama-c -lannot-list`. It also returns the labelized C file.

val luncov_init : (module Types.CONFIG) -> [> `C ] Sc_sys.File.TYPES.file list -> Sc_sys.Process.command Lwt.t

Returns the command line for initializing the labels file with `luncov` on the project in argument. /!\ The files of the project must be the files generated by `lannot`.

val luncov_cmd : (module Types.CONFIG) -> ?main_name:string -> framac_tools:string list -> extra_args:string list -> [> `C ] Sc_sys.File.TYPES.file list -> Sc_sys.Process.command Lwt.t

Returns the command line for starting `luncov` on the project in argument. /!\ The files of the project must be the files generated by `lannot`.

val eacsl_cmd : (module Types.CONFIG) -> ?main_name:string -> output:string -> [> `C ] Sc_sys.File.TYPES.file list -> Sc_sys.Process.command Lwt.t

val check_option : errlog:Sc_sys.Process.TYPES.line_logger -> string -> bool Lwt.t

Returns an lwt promise stating if frama-c offers the given option; logs stderr using the given logger.

val luncov_installed : unit -> bool Lwt.t

Returns an lwt promise stating if luncov is installed

val eacsl_installed : unit -> bool Lwt.t

Returns an lwt promise stating if e-acsl is installed

module Arch : sig ... end

frama-C compatible architectures

val remove_framac_builtins_from_file : [> `C ] Sc_sys.File.TYPES.file -> unit

Removes the "__fc_stuff" added by Frama-C and replaces them by their original name.