Module Sc_ltest.Printer

val pp_lreplay_results : (Types.lreplay_results * Types.lreplay_display_style) Fmt.t

pp_lreplay_results pretty-prints results of lreplay in a classical way, with first a summary of coverage, and then one label status result per line.

val pp_error : Types.error Fmt.t