Module Sc_ltest.Framac

1. Process command related functions

type funs = [

| `All

| `Only of string list

]

type options

module Cmd : Sc_lib.Ezcmd.T

Command line representation for frama-c

val lannot_cmd : (module Sc_lib.Config.T) -> cover_criterion:string -> annoted_functions:funs -> output_file:string -> [> `C ] Sc_sys.File.TYPES.file -> Lwt_process.command Lwt.t

lannot_cmd annot_kind proj Returns the command line for starting `lannot` with the coverage criterion `annot_kind` and the label file generated. You can get the list of all coverage criterions with `$ frama-c -lannot-list`. It also returns the labelized C file.

val luncov_init : (module Sc_lib.Config.T) -> [> `C ] Sc_sys.File.TYPES.file list -> Lwt_process.command Lwt.t

Returns the command line for initializing the labels file with `luncov` on the project in argument. /!\ The files of the project must be the files generated by `lannot`.

val luncov_cmd : (module Sc_lib.Config.T) -> ?main_name:string -> framac_tools:string list -> extra_args:string list -> [> `C ] Sc_sys.File.TYPES.file list -> Lwt_process.command Lwt.t

Returns the command line for starting `luncov` on the project in argument. /!\ The files of the project must be the files generated by `lannot`.

val eacsl_cmd : (module Sc_lib.Config.T) -> ?main_name:string -> output:string -> [> `C ] Sc_sys.File.TYPES.file list -> Lwt_process.command Lwt.t

val cg2json_cmd : (module Sc_lib.Config.T) -> output:[ `json ] Sc_sys.File.t -> [> `C ] Sc_sys.File.TYPES.file -> Lwt_process.command Lwt.t

Returns the commandline for starting the cg2json plugin. The plugin will write in the output file the call graph starting from the entrypoint specified in the configuration module in argument.

val lannot_installed : unit -> bool Lwt.t

Returns an lwt promise stating if lannot is installed

val luncov_installed : unit -> bool Lwt.t

Returns an lwt promise stating if luncov is installed

val eacsl_installed : unit -> bool Lwt.t

Returns an lwt promise stating if e-acsl is installed

val cg2json_installed : unit -> bool Lwt.t

Returns an lwt promise stating if cg2json is installed

module Arch : sig ... end

frama-C compatible architectures

val remove_framac_builtins_from_file : [> `C ] Sc_sys.File.TYPES.file -> unit

Removes the "__fc_stuff" added by Frama-C and replaces them by their original name.